third_party/blink/web_tests/platform/generic/fast/parser/comment-in-textarea-expected.txt

Test that the HTML parser correctly parses the comment-like string in the textarea.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textAreas.length is 1
FAIL textAreas[0].innerHTML should be &lt;!-- &lt;/textarea&gt; --&gt; This should be part of the textarea. Was &lt;!-- .
PASS successfullyParsed is true

TEST COMPLETE

 --> This should be part of the textarea
